+++ ios/Classes/SwiftQonversionFlutterSdkPlugin.swift
@@ -69,6 +69,9 @@ public class SwiftQonversionFlutterSdkPlugin: NSObject, FlutterPlugin {
       qonversionSandwich?.logout()
       return result(nil)
 
+    case "presentCodeRedemptionSheet":
+      return presentCodeRedemptionSheet(result)
+
     default:
       break
     }
@@ -123,6 +126,9 @@ public class SwiftQonversionFlutterSdkPlugin: NSObject, FlutterPlugin {
     case "handleNotification":
       return handleNotification(args, result)
       
+    case "getNotificationCustomPayload":
+      return getNotificationCustomPayload(args, result)
+
     default:
       return result(FlutterMethodNotImplemented)
     }
@@ -283,6 +289,22 @@ public class SwiftQonversionFlutterSdkPlugin: NSObject, FlutterPlugin {
     result(isPushHandled)
   }
   
+  private func getNotificationCustomPayload(_ args: [AnyHashable: Any], _ result: @escaping FlutterResult) {
+    guard let notificationData = args["notificationData"] as? [AnyHashable: Any] else {
+      return result(FlutterError.noData)
+    }
+
+    let customPayload: [AnyHashable: Any]? = qonversionSandwich?.getNotificationCustomPayload(notificationData)
+    result(customPayload?.toJson())
+  }
+
+  private func presentCodeRedemptionSheet(_ result: @escaping FlutterResult) {
+    if #available(iOS 14.0, *) {
+      qonversionSandwich?.presentCodeRedemptionSheet()
+    }
+    result(nil)
+  }
+
   private func getDefaultCompletion(_ result: @escaping FlutterResult) -> BridgeCompletion {
     return { data, error in
       if let error = error {