+++ lib/src/constants.dart
@@ -47,6 +47,8 @@ class Constants {
       'setAppleSearchAdsAttributionEnabled';
   static const mSetNotificationsToken = 'setNotificationsToken';
   static const mHandleNotification = 'handleNotification';
+  static const mGetNotificationCustomPayload = 'getNotificationCustomPayload';
+  static const mPresentCodeRedemptionSheet = 'presentCodeRedemptionSheet';
 
   // Numeric constants
   static const skuDetailsPriceRatio = 1000000;